docs(rust): describe the licenses API in OpenAPI
* Additionally, move the "license" module to agama-lib.
