+++ rust/agama-lib/Cargo.toml
@@ -35,6 +35,8 @@ chrono = { version = "0.4.38", default-features = false, features = [
 home = "0.5.9"
 strum = { version = "0.26.3", features = ["derive"] }
 fs_extra = "1.3.0"
+serde_with = "3.12.0"
+regex = "1.11.1"
 
 [dev-dependencies]
 httpmock = "0.7.0"
