+++ rust/agama-server/src/software.rs
@@ -20,4 +20,3 @@
 
 pub mod web;
 pub use web::{software_service, software_streams};
-mod license;
