+++ rust/agama-server/src/web/docs/software.rs
@@ -37,12 +37,14 @@ impl ApiDocBuilder for SoftwareApiDocBuilder {
             .path_from::<crate::software::web::__path_deregister>()
             .path_from::<crate::software::web::__path_get_config>()
             .path_from::<crate::software::web::__path_get_registration>()
+            .path_from::<crate::software::web::__path_license>()
+            .path_from::<crate::software::web::__path_licenses>()
             .path_from::<crate::software::web::__path_patterns>()
             .path_from::<crate::software::web::__path_probe>()
             .path_from::<crate::software::web::__path_products>()
             .path_from::<crate::software::web::__path_proposal>()
-            .path_from::<crate::software::web::__path_repositories>()
             .path_from::<crate::software::web::__path_register>()
+            .path_from::<crate::software::web::__path_repositories>()
             .path_from::<crate::software::web::__path_set_config>()
             .path_from::<crate::software::web::__path_set_resolvables>()
             .build()
@@ -53,14 +55,17 @@ impl ApiDocBuilder for SoftwareApiDocBuilder {
             .schema_from::<agama_lib::product::Product>()
             .schema_from::<agama_lib::product::RegistrationRequirement>()
             .schema_from::<agama_lib::software::Pattern>()
+            .schema_from::<agama_lib::software::SelectedBy>()
+            .schema_from::<agama_lib::software::model::LanguageTag>()
+            .schema_from::<agama_lib::software::model::License>()
+            .schema_from::<agama_lib::software::model::LicenseContent>()
+            .schema_from::<agama_lib::software::model::RegistrationError>()
             .schema_from::<agama_lib::software::model::RegistrationInfo>()
             .schema_from::<agama_lib::software::model::RegistrationParams>()
-            .schema_from::<agama_lib::software::model::RegistrationError>()
+            .schema_from::<agama_lib::software::model::Repository>()
             .schema_from::<agama_lib::software::model::ResolvableParams>()
             .schema_from::<agama_lib::software::model::ResolvableType>()
-            .schema_from::<agama_lib::software::SelectedBy>()
             .schema_from::<agama_lib::software::model::SoftwareConfig>()
-            .schema_from::<agama_lib::software::model::Repository>()
             .schema_from::<crate::software::web::SoftwareProposal>()
             .schema_from::<crate::web::common::Issue>()
             .build()
